backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
faker
scikit-learn
pandas
anthropic
python-dotenv
